requirements.txt

httpx>=0.27.0
python-dotenv>=1.0.1
feedparser>=6.0.11
beautifulsoup4>=4.12.3
lxml>=5.2.2
pyyaml>=6.0.2
sqlite-utils>=3.36
requests>=2.32.3
requests-oauthlib>=1.3.1
